refactor(hush-cli): extract pkg_cli tests into tests/ subtree
Replace the inline #[cfg(test)] mod tests block with a sibling child
module: pkg_cli/mod.rs now declares `#[cfg(test)] mod tests;` whose body
lives in pkg_cli/tests/. tests/mod.rs holds `use super::*` + the run_cmd
helper and include!()s four de-indented theme files
(scaffold_and_pack.rs, commands.rs, validation.rs, registry.rs) that
carry the 41 tests verbatim, in source order. No test bodies, names, or
assertions changed.
